A wooden box mysteriously appeared on the runway in front of the 707.
The sudden appearance initially surprised the British crew a bit. The
pilot bended a little bit forward to look out of the window, turned his
head to look at the co-pilot and engineer, and said: "Do you think it is
what I think it is?" "Oh absolutely sir, that box must be filled with
lovely green tea," the co-pilot said. The engineer, always coming up
with creative solutions, suggested to the pilot to open his window in
order to try to smell it, which the pilot did. The strong smell of tea
filled the cockpit. It was irresistable to the British crew, and they
spontaneously jumped out of their seats and abandoned the cockpit.